+++ packages/backend/src/module.d.ts
@@ -3,6 +3,5 @@ declare module '@backstage/plugin-auth-backend';
 declare module '@backstage/plugin-auth-backend-module-guest-provider';
 declare module '@backstage/plugin-permission-backend/alpha';
 declare module '@janus-idp/backstage-plugin-matomo-backend/alpha';
-declare module '@appdev-platform/backstage-plugin-report-portal-backend/alpha';
 declare module '@appdev-platform/backstage-plugin-catalog-backend-module-cmdb/alpha';
 declare module '@appdev-platform/backstage-plugin-catalog-backend-module-spaship/alpha';